split interfaces. fixes #60
