+++ api-tests/pom.xml
@@ -3,7 +3,7 @@
          xmlns="http://maven.apache.org/POM/4.0.0"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>selenide-framework</artifactId>
+        <artifactId>sylenium</artifactId>
         <groupId>io.github.symonk</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
