+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>io.github.symonk</groupId>
-    <artifactId>selenide-framework</artifactId>
+    <artifactId>sylenium</artifactId>
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
     <modules>
